+++ webui/tests.py
@@ -1,3 +1,8 @@
 from django.test import TestCase
+from video_screening.py import main
 
 # Create your tests here.
+def test_main():
+    print("Calling main...")
+    main()
+    print("Main called")
